adaptating certificate and sticker image

+++ GMB.Sdk.Core/ToolBox.cs
@@ -486,7 +486,7 @@ namespace GMB.Sdk.Core
         /// <param name="averageScore"></param>
         /// <param name="name"></param>
         /// <returns>certificate</returns>
-        public static Bitmap CreateCertificate(int score1, int score2, int score3, int score4, int score5, decimal averageScore, string name, DateTime orderDate)
+        public static Bitmap CreateCertificate(int score1, int score2, int score3, int score4, int score5, double averageScore, string name, DateTime orderDate)
         {
 
             // Define image size
